replace MapView with new implementation
